+++ Shared/M2CXHelpers.cpp
@@ -8,6 +8,18 @@ License: The MIT License
 #include "pch.h"
 
 #include <Windows.h>
+#include <wrl\client.h>
+#include <wrl\implements.h>
+#include <robuffer.h>
+#include <windows.storage.streams.h>
+
+#include <string>
+
+using Microsoft::WRL::ComPtr;
+using Microsoft::WRL::MakeAndInitialize;
+using Microsoft::WRL::RuntimeClass;
+using Microsoft::WRL::RuntimeClassFlags;
+using Microsoft::WRL::RuntimeClassType;
 
 #include "M2CXHelpers.h"
 
@@ -138,4 +150,167 @@ Platform::Guid M2CreateGuid()
 	GUID guid = { 0 };
 	M2ThrowPlatformExceptionIfFailed(CoCreateGuid(&guid));
 	return Platform::Guid(guid);
+}
+
+// Retrieves the raw pointer from the provided IBuffer object. 
+// Parameters:
+//   Buffer: The IBuffer object you want to retrieve the raw pointer.
+// Return value:
+//   If the function succeeds, the return value is the raw pointer from the 
+//   provided IBuffer object. If the function fails, the return value is 
+//   nullptr.
+// Warning: 
+//   The lifetime of the returned buffer is controlled by the lifetime of the 
+//   buffer object that's passed to this method. When the buffer has been 
+//   released, the pointer becomes invalid and must not be used.
+byte* M2GetPointer(Windows::Storage::Streams::IBuffer^ Buffer)
+{
+	byte* pBuffer = nullptr;
+	ComPtr<Windows::Storage::Streams::IBufferByteAccess> bufferByteAccess;
+	if (SUCCEEDED(ComPtr<IInspectable>(M2GetInspectable(Buffer)).As(
+		&bufferByteAccess)))
+	{	
+		bufferByteAccess->Buffer(&pBuffer);
+	}
+
+	return pBuffer;
+}
+
+class BufferReference : public RuntimeClass<
+	RuntimeClassFlags<RuntimeClassType::WinRtClassicComMix>,
+	ABI::Windows::Storage::Streams::IBuffer,
+	Windows::Storage::Streams::IBufferByteAccess>
+{
+private:
+	UINT32 m_Capacity;
+	UINT32 m_Length;
+	byte* m_Pointer;
+
+public:
+	virtual ~BufferReference()
+	{
+	}
+
+	STDMETHODIMP RuntimeClassInitialize(
+		byte* Pointer, UINT32 Capacity)
+	{
+		m_Capacity = Capacity;
+		m_Length = Capacity;
+		m_Pointer = Pointer;
+		return S_OK;
+	}
+
+	// IBufferByteAccess::Buffer
+	STDMETHODIMP Buffer(byte** value)
+	{
+		*value = m_Pointer;
+		return S_OK;
+	}
+
+	// IBuffer::get_Capacity
+	STDMETHODIMP get_Capacity(UINT32* value)
+	{
+		*value = m_Capacity;
+		return S_OK;
+	}
+
+	// IBuffer::get_Length
+	STDMETHODIMP get_Length(UINT32* value)
+	{
+		*value = m_Length;
+		return S_OK;
+	}
+
+	// IBuffer::put_Length
+	STDMETHODIMP put_Length(UINT32 value)
+	{
+		if (value > m_Capacity)
+			return E_INVALIDARG;
+		m_Length = value;
+		return S_OK;
+	}
+};
+
+// Retrieves the IBuffer object from the provided raw pointer.
+// Parameters:
+//   Pointer: The raw pointer you want to retrieve the IBuffer object.
+//   Capacity: The size of raw pointer you want to retrieve the IBuffer object.
+// Return value:
+//   If the function succeeds, the return value is the IBuffer object from the 
+//   provided raw pointer. If the function fails, the return value is nullptr.
+// Warning: 
+//   The lifetime of the returned IBuffer object is controlled by the lifetime 
+//   of the raw pointer that's passed to this method. When the raw pointer has 
+//   been released, the IBuffer object becomes invalid and must not be used.
+Windows::Storage::Streams::IBuffer^ M2MakeIBuffer(
+	byte* Pointer, 
+	UINT32 Capacity)
+{
+	using Windows::Storage::Streams::IBuffer;
+
+	IBuffer^ buffer = nullptr;
+
+	ComPtr<BufferReference> bufferReference;
+	if (SUCCEEDED(MakeAndInitialize<BufferReference>(
+		&bufferReference, Pointer, Capacity)))
+	{
+		buffer = reinterpret_cast<IBuffer^>(bufferReference.Get());
+	}
+
+	return buffer;
+}
+
+// Converts from the C++/CX string to the UTF-16 string.
+// Parameters:
+//   PlatformString: The C++/CX string you want to convert.
+// Return value:
+//   The return value is the UTF-16 string.
+std::wstring M2MakeUTF16String(Platform::String^ PlatformString)
+{
+	return std::wstring(PlatformString->Data(), PlatformString->Length());
+}
+
+// Converts from the C++/CX string to the UTF-8 string.
+// Parameters:
+//   PlatformString: The C++/CX string you want to convert.
+// Return value:
+//   The return value is the UTF-8 string.
+std::string M2MakeUTF8String(Platform::String^ PlatformString)
+{
+	std::string UTF8String;
+
+	int UTF8StringLength = WideCharToMultiByte(
+		CP_UTF8,
+		0,
+		PlatformString->Data(),
+		static_cast<int>(PlatformString->Length()),
+		nullptr,
+		0,
+		nullptr,
+		nullptr);
+	if (UTF8StringLength > 0)
+	{
+		UTF8String.resize(UTF8StringLength);
+		WideCharToMultiByte(
+			CP_UTF8,
+			0,
+			PlatformString->Data(),
+			static_cast<int>(PlatformString->Length()),
+			&UTF8String[0],
+			UTF8StringLength,
+			nullptr,
+			nullptr);
+	}
+
+	return UTF8String;
+}
+
+// Converts from the UTF-8 string to the C++/CX string.
+// Parameters:
+//   UTF16String: The UTF-16 string you want to convert.
+// Return value:
+//   The return value is the C++/CX string.
+Platform::String^ M2MakeCXString(const std::wstring& UTF16String)
+{
+	return ref new Platform::String(UTF16String.c_str(), UTF16String.size());
 }