+++ Shared/NagisaVersion.h
@@ -12,7 +12,7 @@ License: The MIT License
 
 #define NAGISA_VER_MAJOR 0
 #define NAGISA_VER_MINOR 3
-#define NAGISA_VER_BUILD 23
+#define NAGISA_VER_BUILD 24
 #define NAGISA_VER_REV 0
 #endif